Templates TSX migration
